% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ensembl_version.R
\name{determine_ensembl_version}
\alias{determine_ensembl_version}
\title{Given a list of Ensembl gene IDs, determine which Ensembl version has the highest number of matches.}
\usage{
determine_ensembl_version(ensids)
}
\arguments{
\item{ensids}{character vector of ensemble gene IDs.}
}
\value{
data.frame containing the Ensembl version and the proportion of matching IDs.
}
\description{
Given a list of Ensembl gene IDs, determine which Ensembl version has the highest number of matches.
}
