% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_transcriptome_map.R
\name{get_chr_length}
\alias{get_chr_length}
\title{Estimate the chromosome lengths from the location of the gene with the highest end position.}
\usage{
get_chr_length(ensembl)
}
\arguments{
\item{ensembl}{GRanges object containing the Ensembl GTF.}
}
\value{
numeric vector of chromosome lengths for the autosomes, sex and mitochondrial chromosomes.
}
\description{
Estimate the chromosome lengths from the location of the gene with the highest end position.
}
