% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_transcriptome_map.R
\name{get_ensembl_genes}
\alias{get_ensembl_genes}
\title{Get ensembl GTF as a GRanges object.}
\usage{
get_ensembl_genes(version = NULL)
}
\arguments{
\item{version}{integer that is the Ensembl version to get. Default = NULL which gets the most recent version.}
}
\value{
GRanges object taht is the Ensembl GTF.
}
\description{
Get ensembl GTF as a GRanges object.
}
