% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_transcriptome_map.R
\name{ggtmap}
\alias{ggtmap}
\title{Make a transcriptome map using the ggplot2 framework.}
\usage{
ggtmap(data, color.points = FALSE, cis.points = FALSE,
  cis.radius = 2, cis.color = "#4286f4")
}
\arguments{
\item{data:}{data.frame (or tibble) with the following columns:
      ensembl: (required) character string containing the Ensembl gene ID.
      qtl_chr: (required) character string containing QTL chromsome.
      qtl_pos: (required) floating point number containing the QTL position 
               in Mb.
      qtl_lod: (optional) floating point number containing the LOD score.
      gene_chr:  (optional) character string containing transcript chromosome.
      gene_start: (optional) character string containing transcript start 
                postion in Mb.
      gene_end:  (optional) character string containing transcript end
               position in Mb.
color.points: logical that is TRUE if the points should be colored by LOD.
cis.points: logical that is TRUE if the points should be colored if they
            are with in cis.
cis.radius: numeric value containing the radius in Mb between a gene and a cis-eQTL.
            Optional.
cis.color: color for cis QTL. Optional.}
}
\value{
plot of the QTL and gene location for each gene.
}
\description{
Make a transcriptome map using the ggplot2 framework.
}
