% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herit_conf_int_plot.R
\name{herit_ci_plot}
\alias{herit_ci_plot}
\title{Plot the heritability estimates and their confidene intervals.}
\usage{
herit_ci_plot(h2, title = "")
}
\arguments{
\item{h2}{matrix of heritability simulations with simulations in rows and chromosomes in columns.}
}
\description{
Plot the heritability estimates and their confidene intervals.
}
