% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herit_per_chr_conf_int.R
\name{herit_per_chr_ci}
\alias{herit_per_chr_ci}
\title{Estimate the confidence interval of a heritability estimate.}
\usage{
herit_per_chr_ci(data, K, nsim = 1000)
}
\arguments{
\item{data}{matrix with the phenotype in the first column and the covariate model matrix
in the remaining columns. Sample IDs should be in rownames(data).}

\item{K}{list of per chromosome kinship matrices.}

\item{nsim}{integer indicating the number of simulations to perform. Default = 1000.}
}
\value{
numeric matrix containing heritabilities estimates for the observed data (in row 1) and the simulated data.
}
\description{
Estimate the confidence interval of a heritability estimate.
}
