% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_features.R
\name{plot_genome_features}
\alias{plot_genome_features}
\title{Given a set of genomic features (genes, SNPs, etc.), plot them on a karyogram.}
\usage{
plot_genome_features(features, col)
}
\arguments{
\item{features}{GenomicRanges object containing the feature locations to plot.}

\item{col}{color to use for the features.}
}
\value{
karyogram of the genome with the features plotted.
}
\description{
Given a set of genomic features (genes, SNPs, etc.), plot them on a karyogram.
}
